Move Canvas and Camera into camview.py, fix imports
